Add files via upload
A to-do list is a tool used for organizing and prioritizing tasks that need to be completed. It is a simple yet effective way to keep track of what needs to be done, and to ensure that tasks are completed in a timely and efficient manner.
To-do lists can be created in many different formats, including paper lists, digital lists, and mobile apps. 
Overall, a to-do list is a simple but effective tool for managing tasks and increasing productivity. Whether you prefer to use a digital or paper list, the key is to create a system that works for you and to use it consistently to stay organized and on track.